requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.8
Flask==3.1.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
pytest-asyncio==0.21.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.34.0
Werkzeug==3.1.3
zipp==3.21.0